# specs/5-cloud-native-event-arch/contracts/todo-events-openapi.yaml
openapi: 3.0.3
info:
  title: Todo Event-Driven API
  description: Event-driven API for Todo management system with Kafka integration
  version: 1.0.0
  contact:
    name: Todo System Team
    email: team@example.com
servers:
  - url: http://localhost:8000
    description: Local development server with Dapr integration
  - url: https://todo-api.example.com
    description: Production server

paths:
  /api/v1/todos:
    post:
      summary: Create a new todo task
      description: Creates a new todo task and publishes a creation event to Kafka
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoRequest'
      responses:
        '201':
          description: Todo created successfully, event published to Kafka
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
      security:
        - bearerAuth: []

    get:
      summary: List todos for user
      description: Retrieves user's todo list from database (synchronous)
      parameters:
        - name: completed
          in: query
          description: Filter by completion status
          required: false
          schema:
            type: boolean
        - name: priority
          in: query
          description: Filter by priority level
          required: false
          schema:
            type: string
            enum: [low, medium, high]
        - name: limit
          in: query
          description: Maximum number of results
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Offset for pagination
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of user's todos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/todos/{id}:
    get:
      summary: Get todo by ID
      description: Retrieves a specific todo by ID (synchronous)
      parameters:
        - name: id
          in: path
          required: true
          description: Todo ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Todo details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update todo
      description: Updates todo details and publishes update event to Kafka
      parameters:
        - name: id
          in: path
          required: true
          description: Todo ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTodoRequest'
      responses:
        '200':
          description: Todo updated successfully, update event published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
      security:
        - bearerAuth: []

    delete:
      summary: Delete todo
      description: Deletes a todo and publishes deletion event to Kafka
      parameters:
        - name: id
          in: path
          required: true
          description: Todo ID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Todo deleted successfully, deletion event published
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - bearerAuth: []

  /api/v1/todos/{id}/complete:
    patch:
      summary: Complete/incomplete todo
      description: Marks a todo as complete or incomplete, publishes status event
      parameters:
        - name: id
          in: path
          required: true
          description: Todo ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                completed:
                  type: boolean
                  description: Whether to mark as completed or not
              required:
                - completed
      responses:
        '200':
          description: Todo status updated, completion event published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
      security:
        - bearerAuth: []

  /api/v1/events/status:
    get:
      summary: Event processing status
      description: Health check for event processing system
      responses:
        '200':
          description: Event system status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventProcessingStatus'
        '503':
          description: Event system unhealthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthCheckResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    CreateTodoRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
          minLength: 1
        description:
          type: string
          maxLength: 1000
        due_date:
          type: string
          format: date-time
        priority:
          type: string
          enum: [low, medium, high]
          default: medium
      required:
        - title

    UpdateTodoRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
          minLength: 1
        description:
          type: string
          maxLength: 1000
        due_date:
          type: string
          format: date-time
        priority:
          type: string
          enum: [low, medium, high]
        completed:
          type: boolean
      additionalProperties: false

    TodoResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        completed:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        due_date:
          type: string
          format: date-time
        priority:
          type: string
          enum: [low, medium, high]
      required:
        - id
        - title
        - completed
        - created_at
        - updated_at

    TodoListResponse:
      type: object
      properties:
        todos:
          type: array
          items:
            $ref: '#/components/schemas/TodoResponse'
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer

    EventProcessingStatus:
      type: object
      properties:
        kafka_connected:
          type: boolean
        dapr_connected:
          type: boolean
        consumer_lag:
          type: object
          properties:
            todo_events:
              type: integer
            activity_stream:
              type: integer
        last_processed_event:
          type: string
          format: date-time
        health:
          type: string
          enum: [healthy, degraded, unhealthy]

    HealthCheckResponse:
      type: object
      properties:
        status:
          type: string
          enum: [pass, fail, warn]
        version:
          type: string
        release_id:
          type: string
        service_id:
          type: string
        service_name:
          type: string
        checks:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: '#/components/schemas/HealthCheckDetail'

    HealthCheckDetail:
      type: object
      properties:
        status:
          type: string
          enum: [pass, fail, warn]
        componentType:
          type: string
        observedValue:
          type: string
        observedUnit:
          type: string
        time:
          type: string
          format: date-time
        output:
          type: string

  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              message:
                type: string

    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: unauthorized
              message:
                type: string
                example: Authentication token required

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: not_found
              message:
                type: string
                example: Resource not found

    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: validation_error
              details:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                    message:
                      type: string